monarudo: initial bringup for droid dna

+++ arch/arm/mach-msm/include/mach/board.h
@@ -543,7 +543,7 @@ struct msm_panel_common_pdata {
 	unsigned num_mdp_clk;
 	int *mdp_core_clk_table;
 	u32 mdp_max_clk;
-	u32 mdp_max_bw;
+	u64 mdp_max_bw;
 	u32 mdp_bw_ab_factor;
 	u32 mdp_bw_ib_factor;
 #ifdef CONFIG_MSM_BUS_SCALING
